Add files via upload
This project is a Netflix homepage clone built using HTML and CSS, designed to replicate the look and feel of Netflix’s landing page. It features a dark theme, hover effects, and a simple, clean UI. The layout is structured using HTML, while CSS is used for styling, flexbox, and grid for responsiveness. The homepage includes a navigation bar, body section and a footer, similar to the original Netflix design. This project is beginner-friendly and helps in understanding front-end development concepts, including layout structuring, and UI/UX design principles.